pyproject.toml
[project]
name = "humana-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "chroma>=0.2.0",
    "googlemaps>=4.10.0",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "openai>=1.76.2",
    "optuna>=4.3.0",
    "pypdf>=5.4.0",
    "pytest>=8.3.5",
    "pytest-xdist>=3.6.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]

[tool.pytest.ini_options]
testpaths  = ["tests"]
pythonpath = ["src"]

[tool.poetry.dev-dependencies]
pypdf = "^5.0"
pytest = "^8.0"
